moves the code generation part into the init stage
This commits deals with the goal of getting all the code generation
tasks done at the init stage in order to place these code artifacts onto
an external volume shared across the whole cluster. Benefit of doing so
is to maintain a 1:1 relationshop between currently running database and
the code running at the application server. Drawback is that due to the
tight coupling of ORM code that has been generated and the vendor
provided bundles this might result in breakage especially in rolling
upgrades where database exists and code needs to be updated.

FIXME: We need further think about which kind of update scenario we want
to support: (a) either incrementally, or (b) rebuilding infrastructure
anew on each update.