make splitting yves/zed within one docker image possible
Enable yves/zed vhosts and php-fpm config  dynamicaly via /entrypoint.sh
This way we still can make use of monit and no changes on the startup
process is required. Also there are just small changes on the deployment
process required. This commit contains both, the changes on the deployment
process and the required functionality in /entrypoint.sh

Note: yves and zed are now both listening on port 80! This is required to
fix the cart/login links in the yves html banner.
This is the reason for the new ENV vars

* PUBLIC_YVES_DOMAIN
* PUBLIC_ZED_DOMAIN

Testing is still required.