Make onbuild trigger independently runnable
This commits further splits the ONBUILD trigger insofar that each layer
(base, dep, shop) got their own COPY and RUN command. This makes each
layer more independant from each other. Reason for this is to reduce
build times while developing a shop. So the layer are order from general
to particular interest. If only the source code changes, the rebuild of
the image will only take place in the 3rd layer, the shop build layer.
The idea is that changes on the foundation are not that often as changes
on the code base. This commits tries to reduce the rebuild needs for the
lower layers.