Remove APPLICATION_ENV as build time option
This commit introduces the premise to use a single image for both
production and development. This means that even the production
containers will have dev dependencies included. Primary reason for this
is the requirement for dev/test/prod parity to ensure the containers
behave the same in all stages and in all environments. Tradeoff for
this premise is again larger effective images. During runtime the
behaviour of the Spryker Application can be controlled by setting
`APPLICATION_ENV` which accepts either `development` or `production`. If
you use the `./docker/run` script this variables will be set
automatically.