Add flag to skip cleanup during build
For debugging purposes it might be useful to debug the build within a
running container instead of rebuilding the images all over again. For
this very use case, set `SKIP_CLEANUP=true` and start the build. At the
point where this build fails, commit the build container as new image
and create a container from this commit manually and start debugging.
Its important to never ever release such intermediate images into the
wild!!! Keep this in mind!